release 2023.10 (#198)
* Update install_deps.sh

* Update install_mingw64_deps.sh

* win32 faust

* Update CMakeLists.txt

* Update CMakeLists.txt

* Update FindRAGEL.cmake

* Update install_mingw64_deps.sh

* Update CMakeLists.txt

* WITH_VEROVIO=OFF update

* verovio link fix

* Update FindLibSamplerate.cmake

* default server

* Update make_bundle.sh

* Update s_midi_coremidi.c

* update midi chan maxvalue

* Update ceammc.synth-help.pd

* Update build_options.cmake

* test fix

* editor updates

* Update ceammc_editor_object.cpp

* linux updates

* Update FindLibMagic.cmake

* cppcheck rewrite

* remove dead code

* Update test_platform.cpp

* Update tcl_autogen.cmake

* Update CMakeLists.txt

* faust updates

* Update hoa_process.cpp

* Update CMakeLists.txt

* Update CMakeLists.txt

* Update config_single.sh

* recho max time update

* active property fix

* Update ceammc_dnd-plugin.tcl

* Update ceammc_dnd-plugin.tcl

* tl.timeline output format added

* doc updates

* timeline updates

* new timeline parser

* updates

* updates

* syntax update

* ragel updates

* ragel updates

* ragel updates

* tl updates

* tempo updates

* duration division

* unused include

* Update ui_preset.cpp

* Update ui_gain.cpp

* level meter update

* ui.incdec min/max properties added

* display_only property added

* env.asr2~ stereo env added

* Update eclass.cpp

* ui bind parser added

* ui cleanup

* class renaming

* syntax update

* Update ui_single_value.cpp

* Update test_parser_ui_bind.cpp

* Update tl_baseobject.cpp

* Update ceammc_proxy.h

* Update parser_ui_bind.h

* ui incdec bind updates

* Update ui_incdec.cpp

* using templates

* ui_bang bind added

* category update

* ui bind updates

* bind param operators added

* bind updates

* symbols update

* move file

* key bindings added

* key updates

* tcl: tablelist added

* tablelist

* Update argcheck2.h

* tableeditor added

* adding const

* tableeditor updates

* preset editor for float

* preset symbol editor

* preset list prepare

* tableeditor size

* Update ceammc_tableeditor.tcl

* timeline updates

* artnet.send cleanup

* doc updates

* doc update

* doc updates

* doc updates

* 3d.decoder help added

* hoa.3d.exchanger~ doc added

* Update ceammc_lib_dumpjson.cpp

* doc: hoa.2d.exchanger~ added

* doc: hoa.3d.scope~ added

* doc: format

* doc: hoa.map.ui added

* doc updates

* Update hoa.map-help.pd

* preset interpolation added

* Update CMakeLists.txt

* hoa: 2d.scope~ doc

* Create hoa.2d.rotate~-help.pd

* recomposer updates

* hoa help updates

* Create hoa.2d.projector~-help.pd

* hoa: 2d meter example added

* Update hoa.2d.meter~-help.pd

* Create hoa.2d.map~-help.pd

* hoa updates

* more hoa updates

* Update hoa.2d.space-help.pd

* Update hoa.3d.exchanger~-help.pd

* basic hoa.3d.meter

* meter updates

* Update hoa.3d.decoder~-help.pd

* new hoa files added

* hoa: 3d.scope updates

* Update hoa.3d.scope~-help.pd

* Update hoa_3d_meter.cpp

* Update hoa.3d.wider~-help.pd

* Update hoa.3d.map~-help.pd

* hoa map updates

* Update hoa_3d_meter.cpp

* hoa doc updates

* Update hoa.3d.meter~-help.pd

* hoa doc update

* Update hoa_3d_meter.cpp

* hoa doc updates

* preset fixes

* hoa doc

* more hoa doc updates

* tutorial hoa

* process 3d support

* hoa tutorial updates

* hoa: tutorial updates

* Update hoa_process_props.cpp

* hoa.process bang method

* hoa: tuto 9 added

* Update Tutorial_8.pd

* hoa doc tuto 10

* hoa: fx updates

* Update hoa.fx.freeverb~-help.pd

* Update hoa.fx.gain~-help.pd

* Update hoa.fx.ringmod~-help.pd

* Update hoa.fx.mirror~-help.pd

* hoa fx updates

* Update hoa.syn.grain~-help.pd

* Update hoa.syn.grain~-help.pd

* Update hoa.fx.freeverb~-help.pd

* Update hoa.syn.ringmod~-help.pd

* Update hoa.fx.convolve~-help.pd

* basic convolution added

* Update base_convolve_tilde.cpp

* update for 64 bits

* Update base_convolve_tilde.cpp

* ui updates

* hoa doc updates

* convolve~ added

* ui updates

* env.asr2~ doc added

* hoa doc updates

* hoa.3d.meter~ added

* hoa doc update

* hoa convolve updates

* ui.cmeter~ added

* doc update

* doc cleanup

* hoa.map.ui updates

* Update hoa-help.pd

* Update hoa.fx.wider~.pd

* hoa update

* casio midi proto added

* obsolete object removed

* tl updates

* doc format and updates

* pddoc updates

* bimap updates

* doc update

* doc update

* doc updates

* Update CHANGELOG.md

* include fix

* internal property fix

* Update live_capture.h

* doc updates

* Update live.capture~.pddoc

* doc updates

* l10n updates

* github address update

* Update ceammc_tableeditor.tcl

* Update CMakeLists.txt

* doc update

* cmake updates

* ir added

* Update base_convolve_tilde.cpp

* Update ui_bind_object.h

* Update parser_ui_bind.h

* Update ui_bind_object.h

* Update ui_circular_meter.cpp

* Create patch_faust.sh

* doc update

* Update preset_base.cpp

* double: fix

* double: fix

* cmake: win def fix

* Update CMakeLists.txt

* Update test_coreaudio.cpp

* 64: fix

* linux: opus check

* linux fix

* linux: snap fix

* Update CMakeLists.txt

* doc update

* rhvoice done messages

* doc update

* Update array_ltcplay_tilde.cpp

* libltc 1.3.1 fix

* linux: verovio fix

* snap updates

* Update proto_ltc_in_tilde.cpp

* Update version.cmake

* release texts update

* update

* win fixes

* Update FindLTC.cmake

* Update fix_dll_deps.py

* win: ltc configs

* win: faust updates

* release texts update

* Update make_lib_win.sh

+++ ceammc/ext/src/lang/lua_interp.cpp
@@ -12,7 +12,7 @@
  * this file belongs to.
  *****************************************************************************/
 #include "lua_interp.h"
-#include "fmt/format.h"
+#include "fmt/core.h"
 #include "lua_func.h"
 #include "lua_stack_guard.h"
 
@@ -283,7 +283,7 @@ end
             case LUA_CMD_NOP: // ignore silently
                 break;
             default:
-                error(fmt::format("unsupported command: {}", cmd.cmd));
+                error(fmt::format("unsupported command: {}", (int)cmd.cmd));
                 break;
             }
         } catch (std::exception& e) {
